Circuits + Circuit Boards (item)
+ small bugfix (making Machine public)
* disabled registry logging
+ circuits for machines

+++ src/items/ItemStack.java
@@ -1,3 +1,3 @@
 package items;
 
-public record ItemStack(Item item, int quantity) {}
+public record ItemStack(Item item, double quantity) {}
